@@ -10,123 +10,157 @@
 // content model, add the entry to this list, and ensure you loop in the
 // content and/or content strategy team for review.
 
-module.exports = [
-  '2fa',
+export default [
+  '2FA',
+  'Access management',
+  'Accounts',
   'Action development',
   'Actions',
+  'Actions Runner Controller',
   'Administrator',
   'Advanced Security',
+  'AI',
+  'Alerts',
   'Amazon ECS',
+  'Analytics',
   'Android',
   'Ant',
-  'Analytics',
   'API',
   'Auditing',
   'Authentication',
   'Azure App Service',
+  'Azure Kubernetes Service',
   'Azure Pipelines',
+  'Azure Static Web Apps',
   'Backups',
+  'Billing',
+  'C/C++',
+  'C#',
+  'Cancellation',
   'CD',
   'CI',
   'CircleCI',
+  'CLI',
   'Clustering',
+  'Code scanning',
+  'CodeQL',
+  'Codespaces',
+  'Collaboration',
+  'Community',
+  'Copilot',
   'Contacting sponsors',
   'Containers',
-  'Code scanning',
-  'Codespaces',
+  'CVEs',
+  'Dependabot',
+  'Dependencies',
+  'Dependency graph',
+  'Dependency review',
+  'Deployment',
+  'Desktop',
+  'Developer',
+  'Device verification',
+  'Discounts',
   'Discussions',
   'Docker',
-  'Developer',
-  'Events',
+  'Downgrades',
+  'Early access',
+  'Education',
+  'Enterprise',
   'Errors',
+  'Events',
+  'Forks',
   'Fundamentals',
   'Git',
+  'GitHub Apps',
   'GitHub Connect',
+  'GitHub search',
+  'GitHub',
   'GitLab',
   'Google Kubernetes Engine',
   'Gradle',
   'GraphQL',
+  'Health',
   'High availability',
+  'Identity',
   'Infrastructure',
+  'Integration',
   'iOS',
+  'Issues',
   'Java',
   'JavaScript',
   'Jenkins',
+  'Kotlin',
+  'Legal',
+  'LFS',
   'Licensing',
   'Logging',
+  'Maintenance',
+  'Marketplace',
   'Maven',
   'Migration',
+  'Mobile',
   'Monitoring',
   'Networking',
   'Node',
+  'Nonprofits',
+  'Notifications',
+  'OAuth apps',
   'Open Source',
   'Organizations',
   'Packages',
   'Packaging',
-  'Performance',
-  'Powershell',
-  'Pre-receive hooks',
-  'Privacy',
-  'Project management',
-  'Publishing',
-  'Python',
-  'Ruby',
-  'Secret scanning',
-  'Security',
-  'Sponsors payments',
-  'Sponsors profile',
-  'Storage',
-  'Support',
-  'Swift',
-  'Travis CI',
-  'Upgrades',
-  'User account',
-  'Webhooks',
-  'Workflows',
-  'Access management',
-  'Accounts',
-  'Billing',
-  'CLI',
-  'Codespaces',
-  'Community',
-  'Desktop',
-  'Device verification',
-  'Early access',
-  'Enterprise',
-  'GitHub',
-  'GitHub Apps',
-  'GitHub search',
-  'Identity',
-  'Issues',
-  'Jobs',
-  'Legal',
-  'LFS',
-  'Maintenance',
-  'Marketplace',
-  'Mobile',
-  'Notifications',
-  'OAuth Apps',
   'Pages',
+  'Performance',
   'Permissions',
   'Policies',
   'Policy',
+  'PowerShell',
+  'Pre-receive hooks',
+  'Privacy',
   'Profile',
   'Profiles',
+  'Project management',
   'Projects',
+  'Publishing',
   'Pull requests',
+  'Python',
+  'Receipts',
   'Repositories',
+  'REST',
+  'Ruby',
+  'SARIF',
+  'Secret scanning',
+  'Secret store',
+  'Security advisories',
+  'Security overview',
+  'Security policies',
+  'Security updates',
   'Security',
-  'Support',
   'Set up',
+  'Spending limits',
+  'Sponsors payments',
+  'Sponsors profile',
+  'Sponsors',
   'SSH',
   'SSO',
+  'Storage',
+  'Support',
+  'Swift',
   'Teams',
+  'Transcripts',
+  'Travis CI',
+  'Trials',
   'Troubleshooting',
   'Upgrades',
+  'User account',
   'Usernames',
+  'Version updates',
   'Visual Studio Code',
-  'Xamarin',
-  'Xamarin.iOS',
+  'Vulnerabilities',
+  'Webhooks',
+  'Workflows',
   'Xamarin.Android',
-  'Xcode'
+  'Xamarin.iOS',
+  'Xamarin',
+  'Xcode',
 ]
